+++ src/frontend/apps/impress/src/features/docs/doc-management/types.tsx
@@ -94,6 +94,7 @@ export interface Doc {
     partial_update: boolean;
     restore: boolean;
     retrieve: boolean;
+    search: boolean;
     update: boolean;
     versions_destroy: boolean;
     versions_list: boolean;